Move writeJUnitOutput to output.go
We have a few other functions in output.go for writing reports in CSV or
JSON. Let's move the JUnit output function there too to simplify
main.go.
